refactor: extract guards from state-machine.ts to guards.ts
